barmaid/requirements.txt
aiofiles==0.8.0
discord.py==2.0.0
